requirements.txt
click===8.1.3
PyGithub===1.55
python-gitlab===3.9.0